Use temp fixtures in all snapshot-workflow tests
