+++ test/snapshot-removal/helpers/macros.js
@@ -4,15 +4,6 @@ const path = require('path');
 const tempy = require('tempy');
 const fse = require('fs-extra');
 
-const withTemporaryFixture = cwd => async (t, implementation, ...args) => {
-	await tempy.directory.task(async temporary => {
-		await fse.copy(cwd, temporary);
-		await implementation(t, temporary, ...args);
-	});
-};
-
-module.exports.withTemporaryFixture = withTemporaryFixture;
-
 async function testSnapshotPruning(t, {
 	cwd,
 	env,
