+++ test/snapshot-workflow/fixtures/first-run/test.js
@@ -1,4 +1,4 @@
-const test = require('ava');
+const test = require(process.env.AVA_PATH); // This fixture is copied to a temporary directory, so require AVA through its configured path.
 
 test('foo', t => {
 	t.snapshot({foo: 'one'});
